pyproject.toml
[project]
name = "mesugaki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.7.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]
